+++ src/main/gen/icu/windea/pls/script/psi/ParadoxScriptTypes.java
@@ -5,7 +5,6 @@ import com.intellij.psi.tree.IElementType;
 import com.intellij.psi.PsiElement;
 import com.intellij.lang.ASTNode;
 import icu.windea.pls.script.psi.impl.*;
-import icu.windea.pls.script.psi.impl.*;
 
 public interface ParadoxScriptTypes {