+++ src/main/gen/icu/windea/pls/script/psi/ParadoxScriptVariableReference.java
@@ -2,8 +2,6 @@
 package icu.windea.pls.script.psi;
 
 import java.util.List;
-
-import icu.windea.pls.script.reference.*;
 import org.jetbrains.annotations.*;
 import com.intellij.psi.PsiElement;
 import icu.windea.pls.script.reference.ParadoxScriptVariablePsiReference;