+++ src/main/gen/icu/windea/pls/script/psi/impl/ParadoxScriptPropertyImpl.java
@@ -2,8 +2,6 @@
 package icu.windea.pls.script.psi.impl;
 
 import java.util.List;
-
-import icu.windea.pls.script.psi.*;
 import org.jetbrains.annotations.*;
 import com.intellij.lang.ASTNode;
 import com.intellij.psi.PsiElement;
@@ -93,6 +91,11 @@ public class ParadoxScriptPropertyImpl extends ParadoxScriptStubElementImpl<Para
   @Nullable
   public String getTruncatedValue() {
     return ParadoxScriptPsiImplUtil.getTruncatedValue(this);
+  }
+
+  @Override
+  public int getDepth() {
+    return ParadoxScriptPsiImplUtil.getDepth(this);
   }
 
   @Override