+++ src/main/kotlin/icu/windea/pls/script/psi/ParadoxScriptPropertyStubElementType.kt
@@ -27,8 +27,8 @@ class ParadoxScriptPropertyStubElementType : IStubElementType<ParadoxScriptPrope
 	
 	override fun createStub(psi: ParadoxScriptProperty, parentStub: StubElement<*>): ParadoxScriptPropertyStub {
 		//这里使用scriptProperty.paradoxDefinitionInfo.name而非scriptProperty.name
-		val definition = psi.paradoxDefinitionInfoNoCheck
-		return ParadoxScriptPropertyStubImpl(parentStub, definition?.name ?: "@", definition?.type?.name ?: "")
+		val definition = psi.paradoxDefinitionInfo
+		return ParadoxScriptPropertyStubImpl(parentStub, definition?.name ?: "@", definition?.type ?: "")
 	}
 	
 	//override fun createStub(tree: LighterAST, node: LighterASTNode, parentStub: StubElement<*>): ParadoxScriptPropertyStub {