+++ CHANGELOG.md
@@ -86,6 +86,7 @@
 * [X] 初步实现`CwtConfigResolver`
 * [X] 实现`CwtConfigGroupsProvider`
 * [X] 更换包名
+* [X] 支持基于cwt配置文件识别脚本文件中的定义（definition），并在gutterIcon、documentation等地方中显示
 * [ ] 实现基于`cwtConfigGroupsCache`的更加具体的规则
 * [ ] 规则文件加载需要一定时间 - 可能需要特殊处理